+++ home.admin/config.scripts/bonus.chantools.sh
@@ -17,7 +17,7 @@ fi
 # command info
 if [ $# -eq 0 ] || [ "$1" = "-h" ] || [ "$1" = "-help" ]; then
  echo "Channel Tools install script"
- echo "/home/admin/config.scrips/bonus.chantools.sh on|off|menu"
+ echo "/home/admin/config.scripts/bonus.chantools.sh on|off|menu"
  echo "Installs the version $pinnedVersion by default."
  exit 1
 fi
@@ -29,12 +29,12 @@ fi
 
 # show info menu
 if [ "$1" = "menu" ]; then
-  dialog --title " Channel Tools " --msgbox "
-Channel Tools is a command line tool.
-Type: 'sudo su - bitcoin' in the command line to switch to the bitcoin user.
-Then see 'chantools' for the options.
+  dialog --title " Channel Tools ${pinnedVersion} " --msgbox "\n
+Channel Tools is a command line tool to rescue locked funds.\n\n
+On terminal use command 'chantools' and follow instructions.\n\n
 Usage: https://github.com/guggero/chantools/blob/master/README.md
-" 10 75
+" 11 75
+  clear
   exit 0
 fi