+++ CHANGES.md
@@ -5,6 +5,7 @@
 - New: Raspberry Pi OS Base Image 64-bit (August 2020)
 - New: Build SD card Image with parameters & FatPack [details](https://github.com/rootzoll/raspiblitz/pull/2044)
 - New: Lightning Terminal 0.4.1-alpha (Loop, Pool & Faraday UI Bundle) [details](https://github.com/lightninglabs/lightning-terminal#lightning-terminal-lit)
+- New: chantools 0.8.2 [details](https://github.com/guggero/chantools/blob/master/README.md)
 - New: Telegraf metrics (experimental) [details](https://github.com/rootzoll/raspiblitz/issues/1369)
 - New: Download whitepaper from blockchain [details](https://github.com/rootzoll/raspiblitz/pull/2017)
 - Update: bitcoin-core version 0.21.0-beta [details](https://github.com/bitcoin/bitcoin/blob/master/doc/release-notes/release-notes-0.21.0.md)